# Allow parameter routes to end with a dot (/foo/:id.json)