@ready4god2513 2018-04-10
@vishr Any thoughts on this?  We've had to add some unpleasant changes in our middleware and it would be great for the echo router to support this pattern out of the box.