@vishr 2018-04-10
The caveat I see is that you can't have param with a `.`. What do you guys think about it?