@Carpetsmoker 2018-04-10
> The caveat I see is that you can't have a param with `.`. What do you guys think about it?

Yeah, that's true. I don't know if anyone actually needs/does this though?

*Any* change to this will be backwards-incompatible. For example someone mentioned routes in the form of `/x/:{id}.json` in the other issue, but this will mean routes with `{` and `}` will no longer work, so that's not compatible either (and also a bit harder to implement).