@ansel1 2018-04-10
It won’t affect us. Periods in param names are probably rare.  

If the allowed chars for param names has been well documented before now, I would not make a breaking change. If it hasn’t been, then this was always the intended behavior, and this is just fixing a bug ;)